colocate cio utils into folder
