+++ src/components/survey/survey-reducer.ts
@@ -2,7 +2,7 @@ import sortingHatData from '@/data/sorting-hat'
 import {CIOSubscriber} from '@/hooks/use-cio'
 import {track} from '@/utils/analytics'
 import {isEmpty} from 'lodash'
-import {cioIdentify} from '@/utils/cio-identify'
+import {cioIdentify} from '@/utils/cio/cio-identify'
 
 const DEFAULT_FIRST_QUESTION = `biggest_path`
 const DEFAULT_FINAL_QUESTION = `thanks`
