+++ src/pages/api/cio/identify/[id].ts
@@ -5,6 +5,7 @@ import {CIO_IDENTIFIER_KEY} from '@/config'
 import {ENCODED_CUSTOMER_IO_TRACKING_API_CREDENTIALS} from '@/lib/customer-io'
 import {inngest} from '@/inngest/inngest.server'
 import {SEND_SLACK_MESSAGE_EVENT} from '@/inngest/events/send-slack-message'
+import {reportCioApiError} from '@/utils/cio/report-cio-api-error'
 
 const serverCookie = require('cookie')
 const axios = require('axios')
@@ -60,29 +61,7 @@ const cioIdentify = async (req: NextApiRequest, res: NextApiResponse) => {
       res.setHeader('Set-Cookie', cioCookie)
       res.status(200).json(customer)
     } catch (error: any) {
-      console.log('sending message', {res: error.response})
-      const cioId = error.response.config.url.split('/').pop()
-      await inngest.send({
-        name: SEND_SLACK_MESSAGE_EVENT,
-        data: {
-          messageType: 'error',
-          message: `CustomerIO responded with a ${error.response.status} / ${error.response.statusText} on ${error.response.request.path}`,
-          attachments: [
-            {
-              color: '#d42115',
-              blocks: [
-                {
-                  type: 'section',
-                  text: {
-                    type: 'mrkdwn',
-                    text: `${process.env.CUSTOMER_IO_WORKSPACE_URL}/${cioId}`,
-                  },
-                },
-              ],
-            },
-          ],
-        },
-      })
+      await reportCioApiError(error)
 
       res.status(200).end()
     }
